[server_utils.py] bring back the download_gdrive
In commit b319500d1da848d7d510b683b13533e9e0ae0ff3 the download_gdrive
function was removed from the server_utils, and instead the code of this
function was taken out of it and put in the download_init_csv. However,
get_ml_data also needs the download_gdrive function.

This commit has re-introduced the download_gdrive function and
substituted the code in the download_init_csv so that there is no code
duplication.