[widgets.py] filter on columns in open_csv
the open_csv function was supposed to be able to filter on column names,
but this functionality was actually not working. It never filtered since
it did not use the start and end that were computed before in the code,
but just the full column name list.

It is now changed so that the dataframe actually gets filtered on the
selected column names.