[widgets.py] move the .label from inside function to notebook
In order to be able to test things, the .label or .value or .selected,
etc. need to be called in the notebook, not in the function itself.
Otherwise it is not possible to give it a standard value as input to
check the function.

In this commit the .label is removed from the widgets.open_csv function
and put into notebook 1.