[project, zooniverse_utils] auto login to zooniverse
the argument auto_test is added to the function pp.get_zoo_info(),
pp.set_zoo_info(), zu_utils.connect_zoo_project. The default of this
argument is False, in which case the function works exactly likes before
and asks you to enter your credentials to connect to zooniverse. But you
can pass an argument [username, password] to this function and then you
get automatically logged in.

This feature is added so that the loggin in can be tested via an
automatic test on github.