[project.py] use self.project_name instead of rederiving from self.Project
Resolves issue #86, since now self.project_name is used, where we
correct for capitals and spaces in the project name. It was chosen to
solve the issue in this way and not by renaming the 'Template project'
in the db_starter, since this correction is usefull for any future
projects. By re-using the self.project_name from the class, we need to
apply this correction only once.