Auto tests (#138)
* [widgets.py] move the .label from inside function to notebook

In order to be able to test things, the .label or .value or .selected,
etc. need to be called in the notebook, not in the function itself.
Otherwise it is not possible to give it a standard value as input to
check the function.

In this commit the .label is removed from the widgets.open_csv function
and put into notebook 1.

* [project, zooniverse_utils] auto login to zooniverse

the argument zoo_cred is added to the function pp.get_zoo_info(),
pp.set_zoo_info(), zu_utils.connect_zoo_project and to choose_workflow. The default of this
argument is False, in which case the function works exactly likes before
and asks you to enter your credentials to connect to zooniverse. But you
can pass an argument [username, password] to this function and then you
get automatically logged in.

This feature is added so that the loggin in can be tested via an
automatic test on github.

* Add test key word argument to util functions to enable testing GUI functions

* [project.py] Change methods to match testing requirements

* Remove torch as top-level import

* Remove track from top level import

* Add test kwarg and missing arguments for testing in utils

* Implement changes based on feedback

* Remove debugging print statements
* Replace int with list for image size in train_yolov5
* Where possible, create external function to avoid duplication for
  testing (t3/t4)

* Remove assert False in project.py

* Remove print statements from yolo_utils.py

* Rename server_dict to server_connection

* Ensure best.pt is used for evaluation

* Fix path to best.pt in yolo_utils

* Compatibility fixes

---------

Co-authored-by: Diewertje11 <diewertje@kpnplanet.nl>