[+code-quality.yaml, +.pylintrc] pylint test
add pylint test to kso-utils. This test is run in the container so that
all requirements are installed. Currently, the test is run while only
checking on errors (specified in the .pylintrc), to not get too many
warnings at once. When we are used to this, we can include more
warnings to improve code quality.

Note: currently the container to use is hard coded, but if we then
update the requirements and do not have a pull to the
kso-object-detection yet, we will not have this present here to do the
pylint test. So this need to change in some way: *Or change entire
structure repos *Or just install packages here instead of opening the
container, however, this will have the problem that it cannot see the
requirements from yolov5 and yolov5_tracker which are needed.