add pylint test to kso-utils (#166)
* [+code-quality.yaml, +.pylintrc] pylint test

add pylint test to kso-utils. This test is run in the container so that
all requirements are installed. Currently, the test is run while only
checking on errors (specified in the .pylintrc), to not get too many
warnings at once. When we are used to this, we can include more
warnings to improve code quality.

Note: currently the container to use is hard coded, but if we then
update the requirements and do not have a pull to the
kso-object-detection yet, we will not have this present here to do the
pylint test. So this need to change in some way: *Or change entire
structure repos *Or just install packages here instead of opening the
container, however, this will have the problem that it cannot see the
requirements from yolov5 and yolov5_tracker which are needed.

* fix: all logging-too-many-args errors

We need to specify in the .pylintrc that we use the new logging format
with {} and this we need to actually apply in all the loggin statements.

* remove unused function export_model

* fix: cleaned up the choose_workflows funct

pylint highlighted issues with the duplication of the choose_workflows function so I removed the function from the project.py as we didn't use it in the tutorials and we already have one "choose_workflow" function in zooniverse_utils.

* fix: removed duplicated choose_folder and footage

* fix: removed unnecessary result assigment

* fix: added def Project_name and black

* [project, server_utils, yolo_utils] Implement pylint fixes

* Implement workaround for declaration of variable in project.py

---------

Co-authored-by: Victor Anton <victor@wildlife.ai>
Co-authored-by: Jurie Germishuys <jurie.germishuys@combine.se>