[-.pylintrc,pyproject.toml,yolo_utils] fix pylint
Previous we had all pylint configuration settings in the .pylintrc.
However, the generated-members argument caused a bug due to which we did
not get an error message for non-existing members from another python
file (see issue kso #241).

This seemed to be something that goes wrong in the formatting of the
.pylintrc, since the same generated-members argument does work when the
configuration is given in the pyproject.toml.

So this commit implements the same pylint configureation, but now
without bug, resulting in that it does give all the errors we want for
as far as we know.

The commit also changes the import of cv2 in yolo_utils, since this
caused many errors.